backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
celery[gevent]==5.5.3
azure-identity==1.23.0
httpx==0.28.1
pydantic==2.11.7
redis==6.2.0